+++ Ya/Program/Interface.hs
@@ -89,26 +89,30 @@ instance Mapping Straight Straight Arrow Arrow
 type family Orientation datastructure where
 	Orientation List = Horizontal
 
--- TODO: use `Mapping` instance instead
-class Scrollable datastructure where
-	scroll :: Orientation datastructure -> Automation
-		`TI` Scrolling datastructure item
+class Scrollable morphism datastructure where
+	scroll :: Orientation datastructure -> morphism
 		`TI` Scrolling datastructure item
 		`TI` Optional (Scrolling datastructure item)
 
 -- TODO: try use the fact that `Horizontal` ~ `Boolean`
 -- `Boolean` is `Representative` for `U_I_I LM`
-instance Scrollable List where
-	scroll (That _) = U_I_UU_II_III `yi` \case
+instance Scrollable Statefully' List where
+	scroll (That _) = W_I_I_II `a` U_I_UU_II_III `yi` \case
 		previous@(U_T_I_TT_I (These (Identity x) (U_T_I_TT_I (These (T_'_I (T_TT_I bs)) (T_'_I (List (Yet f fs))))))) -> These
 			(U_T_I_TT_I (These (Identity f) (U_T_I_TT_I (These (T_'_I (List (Yet x (bs `yo` unwrap)))) (T_'_I (T_TT_I / fs `yo` wrap ))))))
 			(Some previous)
 		previous@(_) -> These previous None
-	scroll (This _) = U_I_UU_II_III `yi` \case
+	scroll (This _) = W_I_I_II `a` U_I_UU_II_III `yi` \case
 		previous@(U_T_I_TT_I (These (Identity x) (U_T_I_TT_I (These (T_'_I (List (Yet b bs))) (T_'_I (T_TT_I fs)))))) -> These
 			(U_T_I_TT_I (These (Identity b) (U_T_I_TT_I (These (T_'_I (T_TT_I / bs `yo` R_U_I_T_I)) (T_'_I (List (Yet x (fs `yo` unwrap))))))))
 			(Some previous)
 		previous@(_) -> These previous None
+
+instance {-# OVERLAPS #-} Scrollable Statefully' t => Scrollable Transition t where
+	scroll orient = W_I_II_I `compose` U_I_UU_II_III
+		`compose` fio (\(These x y) -> These y x)
+		`compose` unwrap `compose` unwrap
+		/ scroll @Statefully' @t orient
 
 -- TODO: experimental and highly likely very inefficient
 -- TODO: should we defined with a wrapper since it's not the only possible implementation?