+++ CHANGELOG.md
@@ -608,3 +608,4 @@
 * Define `rwr_fio` operator
 * Define `o_rwr_yio` operator
 * Define `pass` expression
+* Define `Statefully'` pattern